# src/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API de gestión de alumnos",
    "description": "Documento generado automaticamente por swagger-autogen",
    "version": "1.0.0"
  },
  "host": "lzrlgb5t-3000.brs.devtunnels.ms",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "BypassTunnelAlert": {
      "type": "apiKey",
      "in": "header",
      "name": "X-Tunnel-Skip-AntiPhishing-Threshold",
      "description": "Escribe \"true\" para saltar el bloqueo de Dev Tunnels"
    }
  },
  "paths": {
    "/prestamos/": {
      "get": {
        "tags": [
          "Prestamos"
        ],
        "description": "Endpoint para listar todos los prestamos ",
        "parameters": [
          {
            "name": "devuelto",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Prestamos"
        ],
        "description": "Endpoint para registrar un Nuevo Prestamo ",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "libroId": {
                  "example": "any"
                },
                "socioId": {
                  "example": "any"
                },
                "fechaPrestamo": {
                  "example": "any"
                },
                "fechaDevolucion": {
                  "example": "any"
                },
                "devuelto": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/prestamos/{id}": {
      "get": {
        "tags": [
          "Prestamos"
        ],
        "description": "Endpoint para mostrar un prestamo por ID ",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Prestamos"
        ],
        "description": "Endpoint para Actualizar los datos de un Prestamo por ID ",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "fechaDevolucion": {
                  "example": "any"
                },
                "devuelto": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Prestamos"
        ],
        "description": "Endpoint para Eliminar el registro de un Prestamo ",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/libros/": {
      "get": {
        "tags": [
          "Libros"
        ],
        "description": "Obtiene la lista de estudiantes con filtros opcionales",
        "parameters": [
          {
            "name": "autor",
            "in": "query",
            "description": "Filtrar por Autor exacto",
            "type": "string"
          },
          {
            "name": "titulo",
            "in": "query",
            "description": "Filtrar por titulo (insensible a mayúsculas)",
            "type": "string"
          },
          {
            "name": "categoria",
            "in": "query",
            "description": "filtrar por categoria",
            "type": "string"
          },
          {
            "name": "disponible",
            "in": "query",
            "description": "Disponibilidad del Libro (true o false)",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Libros"
        ],
        "summary": "crear un Libro nuevo",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "titulo": {
                  "example": "any"
                },
                "autor": {
                  "example": "any"
                },
                "categoria": {
                  "example": "any"
                },
                "stock": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/libros/{id}": {
      "get": {
        "tags": [
          "Libros"
        ],
        "description": "Obtiene la informacion de un libro buscando por si ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID del libro a buscar"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "Libros"
        ],
        "summary": "actualizar un Libro existente",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID del libro a actualizar"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Datos a actualizar del estudiante",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "titulo": {
                  "type": "string",
                  "example": "La Llorona"
                },
                "autor": {
                  "type": "string",
                  "example": "Miguel de Cervantes"
                },
                "categoria": {
                  "type": "string",
                  "example": "Romance"
                },
                "stock": {
                  "type": "number",
                  "example": 10
                },
                "disponible": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Libros"
        ],
        "summary": "eliminar un Libro",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID del libro a eliminar"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  },
  "security": [
    {
      "BypassTunnelAlert": []
    }
  ]
}